Basics working (no unpacking)

+++ internal/pkg/gitjacker/retriever_test.go
@@ -51,8 +51,12 @@ func (v *vulnerableServer) Addr() string {
 }
 
 func (v *vulnerableServer) Close() error {
-	if err := os.RemoveAll(v.dir); err != nil {
-		return err
+	if os.Getenv("TEST_DEBUG") != "" {
+		fmt.Printf("Input: %s\n", v.dir)
+	} else {
+		if err := os.RemoveAll(v.dir); err != nil {
+			return err
+		}
 	}
 	return v.server.Close()
 }
@@ -100,7 +104,13 @@ func TestRetrieval(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	defer func() { _ = os.RemoveAll(outputDir) }()
+	defer func() {
+		if os.Getenv("TEST_DEBUG") != "" {
+			fmt.Printf("Output: %s\n", outputDir)
+		} else {
+			_ = os.RemoveAll(outputDir)
+		}
+	}()
 
 	if err := New(target, outputDir).Run(); err != nil {
 		t.Fatal(err)
